Annotation update

+++ annotations/nocedit.exe/pseudocode/src/core/cloth.cpp/CClothList_render_FUN_0043c320.cpp
@@ -25,7 +25,7 @@ void __cdecl core_cloth_cpp_CClothList_render_FUN_0043c320(CClothList *this_ptr,
   if (0 < this_ptr->count) {
     do {
       iVar2 = iVar2 + 1;
-      core_cloth_cpp_CCloth_render_FUN_0043bae0(pCVar1->cloths[0],(int)model_ptr);
+      core_cloth_cpp_CCloth_render_FUN_0043bae0(pCVar1->cloths[0],model_ptr);
       pCVar1 = (CClothList *)pCVar1->filenames;
     } while (iVar2 < this_ptr->count);
   }
