+++ annotations/nocedit.exe/pseudocode/src/core/cloth.cpp/CClothList_render_FUN_0043c320.keep.cpp
@@ -18,7 +18,7 @@ void __cdecl core_cloth_cpp_CClothList_render_FUN_0043c320(CClothList *this_ptr,
     }
   }
   for (iVar2 = 0; iVar2 < this_ptr->count; iVar2 = iVar2 + 1) {
-    core_cloth_cpp_CCloth_render_FUN_0043bae0(this_ptr->cloths[iVar2],(int)model_ptr);
+    core_cloth_cpp_CCloth_render_FUN_0043bae0(this_ptr->cloths[iVar2],model_ptr);
   }
   return;
 }
